Revision of the functions to run along the segments to take into account the depolarizations.
Update the functions to addd a counter with a progress bar.

Add a get_version function to give the GA version in order to
automatically know what CCTool-OO Schema use.
Re-work of the function to load the info of the CCTool-OO version and
create a file with a DCSYS Class to know the columns of the different
objects attributes.

Commit before re-work of all the files to use this class to know the
columns rather than the actual values set in stone.